+++ installers/common/MacInstaller.hs
@@ -25,6 +25,7 @@ import           Text.RawString.QQ
 import           System.IO                 (BufferMode (NoBuffering),
                                             hSetBuffering)
 import           System.IO.Error           (IOError, isDoesNotExistError)
+import qualified System.Info
 import           System.Environment        (getEnv)
 import           System.Posix.Files
 import           Turtle                    hiding (e, prefix, stdout)
@@ -151,15 +152,21 @@ sign_cmd "$ABS_PATH/Contents/Resources/app/build/usb_bindings.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/build/HID.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/build/detection.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/keccak/bin/darwin-x64-"*"/keccak.node"
+sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/keccak/bin/darwin-arm64-"*"/keccak.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/keccak/build/Release/addon.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/keccak/prebuilds/darwin-x64/node.napi.node"
+sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/keccak/prebuilds/darwin-arm64/node.napi.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/blake-hash/prebuilds/darwin-x64/node.napi.node"
+sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/blake-hash/prebuilds/darwin-arm64/node.napi.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/blake-hash/bin/darwin-x64-"*"/blake-hash.node"
+sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/blake-hash/bin/darwin-arm64-"*"/blake-hash.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/blake-hash/build/Release/addon.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/blake2/build/Release/binding.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/blake2/bin/darwin-x64-"*"/blake2.node"
+sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/blake2/bin/darwin-arm64-"*"/blake2.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/tiny-secp256k1/build/Release/secp256k1.node"
 sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/tiny-secp256k1/bin/darwin-x64-"*"/tiny-secp256k1.node"
+sign_cmd "$ABS_PATH/Contents/Resources/app/node_modules/tiny-secp256k1/bin/darwin-arm64-"*"/tiny-secp256k1.node"
 
 # Sign the whole component deeply
 sign_cmd "$ABS_PATH"
@@ -213,7 +220,10 @@ buildElectronApp darwinConfig@DarwinConfig{dcAppName, dcAppNameApp} installerCon
 
   let
     formatter :: Format r (Text -> Text -> r)
-    formatter = "../release/darwin-x64/" % s % "-darwin-x64/" % s
+    formatter =
+      if System.Info.arch == "aarch64"
+      then "../release/darwin-arm64/" % s % "-darwin-arm64/" % s
+      else "../release/darwin-x64/" % s % "-darwin-x64/" % s
     pathtoapp :: Text
     pathtoapp = format formatter dcAppName dcAppNameApp
     externalYarn :: [FilePath]
@@ -298,6 +308,8 @@ npmPackage DarwinConfig{dcAppName} = do
   procs "yarn" ["install", "--frozen-lockfile"] empty
   echo "Running electron packager script..."
   export "NODE_ENV" "production"
+  homeDir <- home
+  export "TMPDIR" . tt $ homeDir </> "electron-rebuild-tmp-dir" -- else, new `electron-rebuild` fails with EACCESS
   procs "yarn" ["run", "package", "--", "--name", dcAppName ] empty
   procs "node_modules/.bin/electron-rebuild" ["-w", "usb-detection", "--useCache", "-s"] empty -- <https://github.com/MadLittleMods/node-usb-detection#install-for-electron>
   size <- inproc "du" ["-sh", "release"] empty