+++ installers/overlays/github.nix
@@ -1,0 +1,29 @@
+{ mkDerivation, aeson, base, base-compat, base16-bytestring, binary
+, binary-instances, bytestring, containers, cryptohash-sha1
+, deepseq, deepseq-generics, exceptions, file-embed, hashable
+, hspec, hspec-discover, http-client, http-client-tls
+, http-link-header, http-types, iso8601-time, lib, mtl, network-uri
+, tagged, text, time-compat, tls, transformers, transformers-compat
+, unordered-containers, vector, vector-instances
+}:
+mkDerivation {
+  pname = "github";
+  version = "0.27";
+  sha256 = "5591645d07c0e8c7ab21d6920f3f7bbb10c52b4331d85d264fdc590961d3986e";
+  libraryHaskellDepends = [
+    aeson base base-compat base16-bytestring binary binary-instances
+    bytestring containers cryptohash-sha1 deepseq deepseq-generics
+    exceptions hashable http-client http-client-tls http-link-header
+    http-types iso8601-time mtl network-uri tagged text time-compat tls
+    transformers transformers-compat unordered-containers vector
+    vector-instances
+  ];
+  testHaskellDepends = [
+    aeson base base-compat bytestring file-embed hspec tagged text
+    unordered-containers vector
+  ];
+  testToolDepends = [ hspec-discover ];
+  homepage = "https://github.com/phadej/github";
+  description = "Access to the GitHub API, v3";
+  license = lib.licenses.bsd3;
+}